refactor: getServerMe를 server 전용 모듈로 분리
